Polishing movement, and reposition game objects when out of bounds and other small addtions.

+++ AsteroidsSDL2/Scripts/Vector2D.cpp
@@ -6,7 +6,7 @@ Vector2D::Vector2D()
 	y = 0;
 }
 
-Vector2D::Vector2D(int x, int y)
+Vector2D::Vector2D(float x, float y)
 {
 	this->x = x;
 	this->y = y;
