+++ AsteroidsSDL2/Scripts/Vector2D.h
@@ -3,10 +3,10 @@
 class Vector2D
 {
 public:
-	int x;
-	int y;
+	float x;
+	float y;
 
 	Vector2D();
-	Vector2D(int x, int y);
+	Vector2D(float x, float y);
 private:
 };
